+++ generators/app/inputs/appName.js
@@ -38,7 +38,7 @@
 	module.exports = class AppName {
 		static set($that) {
 			$that.logHintHeader();
-			$that.logHints(hints);
+			$that.logHintsMultiline(hints);
 
 			return $that.prompt(prompts).then($response => {
 				this.onSuccess($that, $response);
@@ -50,6 +50,10 @@
 			$that.appNameCamel       = camelCase($response.appName);
 			$that.appNameKebab       = _.kebabCase($response.appName);
 			$that.appNameUpperPython = upperPythonCase($response.appName);
+			$that.config.set('appName', $that.appName);
+			$that.config.set('appNameCamel', $that.appNameCamel);
+			$that.config.set('appNameKebab', $that.appNameKebab);
+			$that.config.set('appNameUpperPython', $that.appNameUpperPython);
 			$that.log(
 				'The app name in camelCase is :',
 				chalk.hex(colors.get('cyan'))($that.appNameCamel)